requirements.txt
fastapi
uvicorn
torch
torchvision
open3d
pillow
jinja2
